[TS] Disable scrolling to the top of the page when the post modal opens.
