--- ts/twitter/src/lib/components/sidebar-post-button/sign-in-prompt-modal.tsx
@@ -1,37 +1,0 @@
-"use client";
-
-import React from "react";
-import {
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
-  Text,
-} from "@chakra-ui/react";
-
-interface SignInPromptModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-}
-
-export const SignInPromptModal: React.FC<SignInPromptModalProps> = ({
-  isOpen,
-  onClose,
-}) => {
-  return (
-    <Modal isOpen={isOpen} onClose={onClose}>
-      <ModalOverlay />
-      <ModalContent>
-        <ModalHeader>Sign In Required</ModalHeader>
-        <ModalCloseButton />
-        <ModalBody>
-          <Text>You need to sign in to post.</Text>
-        </ModalBody>
-        <ModalFooter></ModalFooter>
-      </ModalContent>
-    </Modal>
-  );
-};
